pizzeria_project/requirements.txt
Django
djangorestframework
celery
eventlet
redis